moved rollbar service into shared package
