+++ apps/passport-server/package.json
@@ -26,6 +26,7 @@
     "@pcd/eddsa-frog-pcd": "0.3.0",
     "@pcd/eddsa-pcd": "0.6.0",
     "@pcd/eddsa-ticket-pcd": "0.6.0",
+    "@pcd/server-shared": "0.0.1",
     "@pcd/email-pcd": "0.6.0",
     "@pcd/message-pcd": "0.1.0",
     "@pcd/passport-crypto": "0.11.0",
