+++ packages/lib/server-shared/src/index.ts
@@ -1,3 +1,1 @@
-export default function (): void {
-  // Package typescript goes here
-}
+export * from "./services/RollbarService";
